pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
